requirements.txt
python-telegram-bot==21.6
gspread==6.1.4
google-auth==2.36.0
openai==1.66.3
pytz==2024.2
